feat: MiniFS v2, ELF64 loader, C用户空间与块设备扩展
- MiniFS v2: 256 inode, 10直接块+1级间接块, 64 KiB文件, 创建/删除/seek/append/truncate
- ELF64 RISC-V 静态加载器, 多页Sv39用户空间, BSS清零, argv/envp栈
- 新增用户crt0/简化libc/C Shell, 以及/bin和/tests下16个独立程序
- Shell支持PATH/引号转义/后台任务/exec/run与< > >>重定向
- CEMU块设备扩展至8 MiB, 修复RV64 LUI符号扩展错误
- 新增tools/mkfs_minifs.py, 从staging rootfs构建磁盘镜像
- 移除旧的user_entry.S内嵌用户态, 改用独立ELF用户程序
- 97/97 CTest测试全部通过

+++ src/instructions.cpp
@@ -422,9 +422,10 @@ std::optional<uint64_t> executeAdd(Cpu& cpu, uint32_t inst) {
 
 std::optional<uint64_t> executeLui(Cpu& cpu, uint32_t inst) {
   auto [rd, rs1, rs2] = unpackInstruction(inst);
-  auto immediate = static_cast<uint64_t>(inst & 0xfffff000);  // Extract the upper 20 bits
+  auto immediate = static_cast<int64_t>(
+      static_cast<int32_t>(inst & 0xfffff000));
   TRACE_LOG("LUI: x", rd , " = ", immediate);
-  cpu.regs[rd] = immediate;
+  cpu.regs[rd] = static_cast<uint64_t>(immediate);
   return cpu.update_pc();
 }
